fixed araqne/issue#1454. expanded offset and limit number range of table command

+++ araqne-logdb/src/main/java/org/araqne/logdb/query/parser/TableParser.java
@@ -126,7 +126,7 @@ public class TableParser extends AbstractQueryCommandParser {
 			to = QueryTokenizer.getDate(options.get("to"));
 
 		if (options.containsKey("offset"))
-			offset = Integer.parseInt(options.get("offset"));
+			offset = Long.parseLong(options.get("offset"));
 
 		if (offset < 0){
 			//throw new QueryParseException("negative-offset", -1);
@@ -137,7 +137,7 @@ public class TableParser extends AbstractQueryCommandParser {
 		}
 
 		if (options.containsKey("limit"))
-			limit = Integer.parseInt(options.get("limit"));
+			limit = Long.parseLong(options.get("limit"));
 
 		if (limit < 0){
 			//throw new QueryParseException("negative-limit", -1);
